Add initial migrations for Django >= 1.7
On Django >= 1.7 you use "./manage.py migrate" rather than syncdb;
this command tries to create tables for all the apps without
migrations before those with migrations.  Creating django-popolo's
tables, however, depends on contenttypes' tables already existing,
and since Django 1.7 contenttypes has migrations.

You can't have an app without migrations depend on one with
migrations, so the only real solution is to add an initial
migration to django-popolo that depends on contenttypes.  It's
a good idea to have migrations for this app anyway in order to
gracefully support future schema changes.

This commit introduces that initial migration (generated with
./manage makemigrations popolo in a Django project that includes
django-popolo) and also removes the entries in .gitignore files
that caused popolo/migrations/ to be ignored.  (I'm not sure why
this was done and the corresponding commit messages don't
explain why.)